+++ RELEASE_NOTES.md
@@ -1,3 +1,7 @@
+### 0.3.0 (2015-06-11)
+* [LanguageExtensions] Require.Positive and Require.ZeroOrGreater method for int type.
+* [EventStore] OpenEventStream reader from specified stream version.
+
 ### 0.2.0 (2015-06-11)
 * [LanguageExtensions] Require.NotEmpty method for collection was added.
 * [LanguageExtensions] IsNullOrEmpty string extension method was added.